+++ API/Controllers/EmployeesController.cs
@@ -1,4 +1,5 @@
 ﻿using API.DAL;
+using API.DAL.Contract;
 using API.Entities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -15,11 +16,11 @@ namespace API.Controllers
     {
         #region ConstructorRegion
 
-        private readonly EmployeesDbContext _employeeDbContext;
+        private readonly IEmployeesRepository _employeesRepository;
 
-        public EmployeesController(EmployeesDbContext dbcontext)
+        public EmployeesController(IEmployeesRepository employeesRepository)
         {
-            _employeeDbContext = dbcontext;
+            _employeesRepository = employeesRepository;
                 
         }
 
@@ -32,22 +33,9 @@ namespace API.Controllers
         /// </summary>
         [HttpPost]
         [Route("RegisterEmployee")]
-        public async Task<bool> RegisterEmployee(Employee employee)
+        public Task<bool> RegisterEmployee(Employee employee)
         {
-            try
-            {
-                employee.CreatedBy = "Admin";
-                employee.CreatedOn = DateTime.Now;
-                employee.Dob = employee.Dob.ToLocalTime();
-                await _employeeDbContext.Employees.AddAsync(employee);
-                var result = await _employeeDbContext.SaveChangesAsync() > 0;
-                return result;
-            }
-            catch (Exception ex)
-            {
-                return false;
-            }
-            
+           return _employeesRepository.RegisterEmployee(employee);      
         }
 
         /// <summary>
@@ -55,17 +43,9 @@ namespace API.Controllers
         /// </summary>
         [HttpGet]
         [Route("GetEmployeeDetails/{name}")]
-        public async Task<Employee> GetEmployeeDetails(string name)
+        public Task<Employee> GetEmployeeDetails(string name)
         {
-            try
-            {
-                return await _employeeDbContext.Employees.FirstAsync(emp => (emp.FirstName == name || emp.LastName == name || (emp.FirstName + ' ' + emp.LastName) == name));
-            }
-            catch (Exception ex)
-            {
-                return null;
-
-            }
+            return _employeesRepository.GetEmployeeDetails(name);
         }
 
         /// <summary>
@@ -73,17 +53,9 @@ namespace API.Controllers
         /// </summary>
         [HttpGet]
         [Route("GetDepartmentMasterData")]
-        public async Task<IEnumerable<Department>> GetDepartmentMasterData()
+        public Task<IEnumerable<Department>> GetDepartmentMasterData()
         {
-            try
-            {
-                return await _employeeDbContext.Departments.ToListAsync();
-            }
-            catch (Exception ex)
-            {
-                return null;
-
-            }
+            return _employeesRepository.GetDepartmentMasterData();
         }
 
         #endregion
